[Refactor] Set m_AttackResetTIme, m_AttackLimitTime Member Variable To Monster So That They Can be Shared Among Classes That Inherits Monster Class

+++ Client2D/Include/Object/Player2D.cpp
@@ -2497,7 +2497,7 @@ void CPlayer2D::PlayerAttackCollisionCallback(const CollisionResult& Result)
 			{
 				CUIDamageFont* DamageFont = ObjectWindow->GetWidgetWindow()->CreateUIWidget<CUIDamageFont>("DamageFont");
 
-				DamageFont->SetDamage(m_AttackAbility);
+				DamageFont->SetDamage((int)m_AttackAbility);
 			}
 		}
 	}
